+++ src/main/java/step2/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package study;
+package step2;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
