+++ src/test/java/step1/SetTest.java
@@ -1,6 +1,5 @@
-package study;
+package step1;
 
-import org.assertj.core.util.Strings;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
