+++ src/test/java/step2/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
